Random renewal of expiring certificates.
Sometime there are a large number of certificates expiring on the same
day (because of bulk creation of domain or glitch in previous renew job
that result in those certificates got issued on the same day). This
could create too many openssl and dehydrated processes and impact
openresty performance. So we try to spread out renewal of those
certificates over a period of 15-30 days before expiry randomly.
Then the number of certificates to be renewed should be spreaded out
uniformly over time.